requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
coverage==7.9.2
idna==3.10
packaging==25.0
paho-mqtt==2.1.0
pip-review==1.3.0
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
